+++ webcomponents/slides/big-img/CHANGELOG.md
@@ -1,6 +1,16 @@
+# 1.1.0 (2020-07-10)
+
+### Features
+
+- add `header` and `footer`
+
+### Refactoring
+
+- remove `custom-actions` and `custom-background` properties because these are only use as attributes
+
 # 1.0.1 (2020-05-11)
 
-### Feat
+### Features
 
 - update Stencil for Gatsby build